requirements.txt
numpy==1.18.0
pandas==1.0.1
Pillow==7.0.0
streamlit==0.58.0
torch==1.4.0
emoji
fastai==1.0.61